+++ src/plugins/piral-breadcrumbs/src/Breadcrumbs.test.tsx
@@ -7,7 +7,6 @@ import { describe, it, expect, vitest, afterEach } from 'vitest';
 import { render, cleanup } from '@testing-library/react';
 import { StateContext } from 'piral-core';
 import { Breadcrumbs } from './Breadcrumbs';
-import { useRouteMatch } from 'react-router';
 
 const MockBcContainer: React.FC<any> = ({ children }) => <div role="container">{children}</div>;
 MockBcContainer.displayName = 'MockBcContainer';
@@ -15,16 +14,7 @@ MockBcContainer.displayName = 'MockBcContainer';
 const MockBcItem: React.FC<any> = ({ children }) => <div role="dialog">{children}</div>;
 MockBcItem.displayName = 'MockBcTile';
 
-vitest.mock('react-router', () => ({
-  useLocation() {
-    return {
-      pathname: '/example',
-    };
-  },
-  useRouteMatch: vitest.fn(() => ({})),
-}));
-
-function createMockContainer(breadcrumbs = {}) {
+function createMockContainer(breadcrumbs = {}, path = '/example') {
   const state = create(() => ({
     components: {
       BreadcrumbsContainer: MockBcContainer,
@@ -46,6 +36,10 @@ function createMockContainer(breadcrumbs = {}) {
       },
       dispatch(update) {
         state.setState(update(state.getState()));
+      },
+      navigation: {
+        path,
+        listen() {},
       },
     } as any,
     api: {} as any,
@@ -104,11 +98,9 @@ describe('Piral-Breadcrumb Container component', () => {
   });
 
   it('dynamic title function replaces wildcard with route param', () => {
-    (useRouteMatch as any).mockReturnValueOnce({ params: { example: 'replacedWildcard' } });
-
     const { context } = createMockContainer({
       example: {
-        matcher: /^\/example$/,
+        matcher: /^\/([a-zA-Z]+)$/,
         settings: {
           path: '/:example*',
           title: ({ path }) => {
@@ -117,7 +109,7 @@ describe('Piral-Breadcrumb Container component', () => {
           parent: '/',
         },
       },
-    });
+    }, '/replacedWildcard');
     const node = render(
       <StateContext.Provider value={context}>
         <Breadcrumbs />
@@ -129,35 +121,7 @@ describe('Piral-Breadcrumb Container component', () => {
     expect(node.getAllByRole('container')[0].innerHTML).toBe('<div role="dialog">/replacedWildcard</div>');
   });
 
-  it('dynamic title function no match in params', () => {
-    (useRouteMatch as any).mockReturnValueOnce({ params: { imNotHere: 'replacedWildcard' } });
-
-    const { context } = createMockContainer({
-      example: {
-        matcher: /^\/example$/,
-        settings: {
-          path: '/:example*',
-          title: ({ path }) => {
-            return path;
-          },
-          parent: '/',
-        },
-      },
-    });
-    const node = render(
-      <StateContext.Provider value={context}>
-        <Breadcrumbs />
-      </StateContext.Provider>,
-    );
-
-    expect(node.getAllByRole('container').length).toBe(1);
-    expect(node.getAllByRole('dialog').length).toBe(1);
-    expect(node.getAllByRole('container')[0].innerHTML).toBe('<div role="dialog">/:example*</div>');
-  });
-
   it('dynamic title function with falsely route param', () => {
-    (useRouteMatch as any).mockReturnValueOnce({ params: { example: false } });
-
     const { context } = createMockContainer({
       example: {
         matcher: /^\/example$/,
@@ -182,8 +146,6 @@ describe('Piral-Breadcrumb Container component', () => {
   });
 
   it('dynamic title function with static title', () => {
-    (useRouteMatch as any).mockReturnValueOnce({ params: { imNotHere: 'replacedWildcard' } });
-
     const { context } = createMockContainer({
       example: {
         matcher: /^\/example$/,
@@ -208,8 +170,6 @@ describe('Piral-Breadcrumb Container component', () => {
   });
 
   it('static title', () => {
-    (useRouteMatch as any).mockReturnValueOnce({ params: {} });
-
     const { context } = createMockContainer({
       example: {
         matcher: /^\/example$/,